chore(eslint): add ESLint configuration file
A new file `.eslintrc.js` has been added to the project. This file contains the configuration for ESLint, a tool used for static code analysis. The configuration sets the environment to support browser and ES2021 syntax. It extends the recommended ESLint rules and includes an override for Node.js environment with specific rules for `.eslintrc.js` and `.eslintrc.cjs` files. The parser options are set to use the latest ECMAScript version. No specific rules have been defined yet.